some quick fixes
pip install fails if PyPDF2 is not already installed, since setup.py
imports pdfmerge for doc and __author__, but pdfmerge in turn imports
PyPDF2. the import now happens if the module works as __main__

also, the command fails if a pdf file is reported as encrypted, even
though no password is set. using default password, which is blank,
fixes these situations.